janium_functions/poll_ulinc_csv/poll_ulinc_csv_director/requirements.txt
sqlalchemy
pymysql
google-cloud-pubsub
requests